+++ pfsimulator/parflow_lib/kinsol_nonlin_solver.c
@@ -170,6 +170,7 @@ int  KINSolInitPC(
   Vector      *saturation = StateSaturation(((State*)current_state));
   Vector      *density = StateDensity(((State*)current_state));
   Vector      *old_pressure = StateOldPressure(((State*)current_state));
+  Vector      *evap_trans = StateEvapTrans(((State*)current_state));
   double dt = StateDt(((State*)current_state));
   double time = StateTime(((State*)current_state));
 
@@ -177,7 +178,7 @@ int  KINSolInitPC(
    * itself */
 
   PFModuleReNewInstanceType(KinsolPCInitInstanceXtraInvoke, precond, (NULL, NULL, NULL, problem_data, NULL,
-                                                                      pressure, old_pressure, saturation, density, dt, time));
+                                                                      pressure, old_pressure, saturation, density, evap_trans, dt, time));
   return(0);
 }
 
@@ -523,7 +524,7 @@ PFModule  *KinsolNonlinSolverInitInstanceXtra(
       instance_xtra->precond =
         PFModuleNewInstanceType(KinsolPCInitInstanceXtraInvoke, public_xtra->precond,
                                 (problem, grid, grid2d, problem_data, temp_data,
-                                 NULL, NULL, NULL, NULL, 0, 0));
+                                 NULL, NULL, NULL, NULL, NULL, 0, 0));
     else
       instance_xtra->precond = NULL;
 
@@ -545,7 +546,7 @@ PFModule  *KinsolNonlinSolverInitInstanceXtra(
       PFModuleReNewInstanceType(KinsolPCInitInstanceXtraInvoke,
                                 instance_xtra->precond,
                                 (problem, grid, grid2d, problem_data, temp_data,
-                                 NULL, NULL, NULL, NULL, 0, 0));
+                                 NULL, NULL, NULL, NULL, NULL, 0, 0));
 
     PFModuleReNewInstanceType(NlFunctionEvalInitInstanceXtraInvoke, instance_xtra->nl_function_eval,
                               (problem, grid, grid2d, temp_data));